backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
httpx
retell
vapi_server_sdk